+++ ec-project-api/Repository/system/StatusRepository.cs
@@ -1,7 +1,7 @@
 using ec_project_api.Interfaces.System;
 using ec_project_api.Models;
 
-public class StatusRepository : Repository<Status, int>, IStatusRepository
+public class StatusRepository : Repository<Status, short>, IStatusRepository
 {
     public StatusRepository(DataContext context) : base(context) { }
 }
